new feature to execute a program for each valid image stream
a wrapper .sh can be used to exec a vlc, ffmpeg or other process to
stream out when we have a valid image. this allows server side to
track when camera is connected or not. will add environment variables
that contain information about the device (MAC, IP, etc) so single
device can stream out data from multiple device in simple manner